less aggressive SPF configuration
We've been using this as a temporary change for over a year since soft
fail is meant to signal that email is likely spam rather than having it
be rejected. This previously helped to cut down on spam but ended up
being too aggressive.

DMARC is what we use to enforce DKIM/SPF but unfortunately isn't very
widely adopted in enforce/quarantine mode.